Add debugging, fix API.
We now use the Gstreamer Debugging API to output debug info. Lots of debug info, controlled by category camsrc.

Also, the API now requires the client to provide the file path for the movie they'd like to receive. And times are in ms since the epoch.

You can still provide a start time that is relative to the current time, handy for testing.